Tidy sequence format
