+++ experimental/dds/tree2/src/feature-libraries/sequence-field/rebase.ts
@@ -39,11 +39,9 @@ import {
 	Changeset,
 	Mark,
 	MarkList,
-	ExistingCellMark,
 	NoopMark,
 	MoveId,
 	Modify,
-	EmptyInputCellMark,
 	NoopMarkType,
 	HasLineage,
 } from "./format";
@@ -58,6 +56,7 @@ import {
 	PairedMarkUpdate,
 } from "./moveEffectTable";
 import { MarkQueue } from "./markQueue";
+import { ExistingCellMark, EmptyInputCellMark } from "./helperTypes";
 
 /**
  * Rebases `change` over `base` assuming they both apply to the same initial state.
